Implement streaming versions of SSE functions
Implement streaming versions of SSE accelerated functions that use
store instructions with a non-temporal hint (so called stream
instructions), resulting in bypassing of the L1 cache. For an amount
of stores/writes that is larger than a value somewhat smaller than
the L1 data cache size, this generally results in higher performance.
Whether to use in streaming or non-streaming version is dynamically
determined (with possible inline optimization of this check).

Signed-off-by: Harm Hanemaaijer <fgenfb@yahoo.com>